Initial coordinator plugin implementation
'files' plugin replaces old backup/restore samples.

The idea (to be documented at some point once it stabilizes) is to
have a set of steps, with each of them being async method
(step_STEPNAME) on the op class.

The steps get as their inputs basically whatever is stored on the
object. Each step's return value is stored as result_STEPNAME if it is
not falsy, and if result is falsy, the operation is considered a
(transient, retryable) failure.

All plugins have also

- their own configuration (stored in coordinator under plugin_config
  key)

- their own (meta)data within backup manifest (not used yet, but with
  same scheme as how configuration is now handled for files)