root_link was not really optional
Added also validation of directory parameters in configuration files (
usually if directory does not exist it is configuration error )

Removed also dangling file_path_filter that seems redundant (for now).
TBD if someday we need to filter things by more than glob, but if we
do, that will come up when it does.