clickhouse: tolerate failure during a restore operation
Two main things were improved to allow for a successful restore after
a failed one :
- Previously created database are dropped before creating them again,
  this also takes care of cleaning partially restored tables.
- Previously created entities are ignored instead of failing when
  trying to create them again.

These two steps are now tested for this behavior.

The main integration test was also modified to test the behavior of
an entire restore after a failure on any step, this greatly increases
the tests duration, the entire suite takes almost 2 minutes now.