coordinator: receive plugin manifest as dict
Current implementation of UploadManifestStep does not take any
advantage of the fact that plugin_manifest is an AstacusModel, but
converts it to dict. This dict then becomes a part of a bigger
BackupManifest model. If a plugin wants to create this dict in some
other way than just calling dict() on the model, this behavior becomes
inconvenient. Therefore, accept an arbitrary dict for upload instead
of AstacusModel.

Another approach could be to incorporate plugin_manifest inside
BackupManifest as an AstacusModel. Then serialization of plugin_manifest
could be customized inside it's model configuration. But, if one day
coordinator decides to configure BackupManifest in a certain way,
plugin manifest models could override that configuration in
unpredictable ways, which would lead to behavior unexpected by
coordinator. Besides, forcing the plugin to use an AstacusModel might
be too restrictive.