fix SnapshotResult.state backward compatibility
This value can be `null`/`None` when received from a node running the
code from before this change:

https://github.com/aiven/astacus/commit/d92e18713b0284cceb62b6bdf0e5cb54ae80c75c

This only happened at the very beginning of a `SnapshotOp`,
when a node can return the value from `SnapshotOp.create_result()`
before the `state` is set in `SnapshotOp.snapshot()`.

This is a partial reversal, which allows the `None` value but stops
generating it.

After all nodes of a cluster are running code newer than the
previously mentioned commit, we will be able to remove the `Optional`
and all the asserts safely.