server: Fix the stand-alone server based on Makefile to actually work
The entry point was somewhat bad, leading app being unset on
e.g. reload (which could be triggered in dev env easily enough).

All of this stuff could be fixed with more recent uvicorn and its
--factory.

This fix was in mst-cassandra branch that was abandoned in June 2021,
and I picked it up from there.