Recreate ClickHouse databases with their original shard/replica/uuid
These values are only recreated if they are part of the backup.

If they are not part of the backup, the uuid will be auto-selected by
ClickHouse (like before), and the shard and replica will be set to
`{shard}` and `{replica}`.

This adds complexity to the `SyncReplicatedDatabaseStep`: we can't
assume anymore that the shard and replica use the default macros,
we have to expand them. And we have to do that separately for each
database, they can have different configurations.