clickhouse: honour sharding when de-duplicating parts
When building a backup manifest, we used to alter how files will be
restored by de-duplicating MergeTree table parts that were identical
on all servers of a cluster.

Now, we do the same thing but instead of de-duplicating across all
servers, we only do that within a single shard.

In general, each Replicated table can have a different sharding
configuration.

In practice, using the Replicated database engine, all tables inside
the same database will have the same sharding configuration.

We use this information to infer the sharding configuration of each
table, and then distribute parts of each tables according to that
configuration.

The restore operation is essentially oblivious to that manipulation.

As long as the cluster where the data is restored has the same
shape (same number of servers with the same sharding configuration
for each database), the data will be restored once per shard.

ClickHouse will replicate the data within each shard, like it already
did when the entire cluster was a single shard.