clickhouse: increase CREATE/DROP DATABASE timeout and concurrency
With larger databases, if the first restoration fails, we need to drop a database, this can take quite some time. Increase the default timeout
to 300sec. Since this can take some time, parallelize the operations,
this can be very beneficial on large clusters.

While we're changing this area, also parallelize the CREATE DATABASE
and increase the timeout, this should speed-up restoration on large
clusters.