Add a node API to get version and features
This allows the coordinator to discover the supported features
of the node and do the correct query.

This is necessary because we can't add new fields on existing ipc
requests without being blocked by the `extra = "forbid"` option
of `NodeRequest` (via `AstacusModel`).

Removing that option would not help on existing deployment
since nodes are already running with the forbid option and
coordinator might be updated before other nodes.