Implement basic partial restore for Cassandra
Base it upon the node replacement for dead nodes (replace_address), but
skip the streaming stage, since we're downloading the data from object
storage. This is not production-ready yet, because the current process
loses all the writes since the last backup when restoring. Incremental
backup and restore to be introduced later.

After downloading the snapshot, stop Cassandra on the node we took the
snapshot of. That node might no longer be in the cluster, we attempt
to match it to configuration nodes the same way we match configuration
nodes to live cluster nodes when creating the backup.

Then start the replacement node with replace_address and
skip_bootstrap_streaming parameters. Wait until it is up, at this point
Cassandra should remove the old node from gossip because we instructed
it to perform a replacement.