snapshot: configure max embed size per group
Allow configuring a different maximum size for embedded file for
each root_glob.

The ipc message is designed to be backward compatible on the node
side and will merge the old format with the new format.

The new format was necessary because root_globs was a list of string,
a list of strings cannot be extended to add attributes for each item.