node: set downloaded files mode to masked 660 instead of 600
This avoids problems astacus and the restored service run with a
different user. If they share a group, the service will then be able to
modify a downloaded file after the restoration has completed.

The 600 permission might have been accidental anyway, it's only
a consequence of trying to atomically create a file using
NamedTemporaryFile.

The effective file mode takes in account the umask, as if the file was
directly created with the specified mode and not created by
NamedTemporaryFile then modified with chmod.