node: add option to chown downloaded files
Even with read/write access for the group and a shared group with
the restored service, there that still some file operation that are
impossible for the service if the file are not owned by the service
(like utime).

Add an option to copy the owner of the download root to each
created file and folder.