Log operation progress on coordinator
When performing cluster operations, the coordinator is waiting for a
certain NodeResult to be returned by the operation status request. But
the logs contain no traces of the NodeResult visible on the coordinator
side. As a result, when an operation takes longer than expected, all
we see in the coordinator logs is a few repeated outgoing status
requests followed by the operation failure. Even if we match the
outgoing requests to the node HTTP server logs, we observe only 200
response codes, but don't see the status inside the request body.

To make operation progress more visible in the coordinator logs, add
intermediate operation result logging in the wait_successful_results
function. This way we can see what the coordinator is waiting for and
try to match this with the logs on the misbehaving node.