Move node features into ipc module
Since Features are imported both from coordinator and node, it follows
that the proper place for this class should be in a common dependency
like the ipc module. Coordinator should not import from node directly.
Move the class and rename it to NodeFeatures to emphasize that these
are the features coordinator is relying on when preparing requests that
all node in the cluster understand.