Remove DownloadKeptBackupManifests step
From the current usage it seems there's no profit in a separate
step for downloading the manifests for backups we've decided to keep
when applying retention rules. We need the full manifests anyway,
because we need to delete dangling hashes, and if we have to look at
backup timestamps in ComputeKeptBackups we end up downloading the same
manifests twice.