coordinator: Make cleanup keep only interesting manifests
The endgame we really want is adding per-node manifests and some sort
of summary manifest to be handled here, not keeping the 'too old'
manifests in memory helps a lot in case of nodes which have already
taken too many backups.

With this, we will load only up to the <amount kept+1> manifests to
memory at most, so having e.g. 70 manifests in object storage and 5
kept means we at most have 6 in memory at a time.

(As seen in production.)