clickhouse: don't SET merge_tree_allow_nullable_key
Setting it was unnecessary and did not help.

The previous change https://github.com/Aiven-Open/astacus/pull/172
was motivated by an incorrect diagnosis of the problem:

This flag is a merge tree setting and not a global setting, so it's
part of the table definition and does not need to be set on the
session.

The real problem was https://github.com/ClickHouse/ClickHouse/issues/54814
which was introduced in ClickHouse 23.7 and fixed in ClickHouse >23.9
and is specific to projections inside a table:

https://github.com/ClickHouse/ClickHouse/pull/52361/commits/781674bdc5b7920c4881614ceb6aa011357e53b7
https://github.com/ClickHouse/ClickHouse/pull/54895/commits/973cd5e972192a2139c20191bf26c75c8a03183f

Another problem is that checkProperties (or checkKeyExpression) is not
called when doing: `ALTER TABLE ... MODIFY SETTING
allow_nullable_key=false` on a table that relied on a nullable key,
but that requires an upstream fix.