cassandra: drop nodes from cluster when they are stopped
cassandra will try to contact these stopped nodes to get hashes, but
that is bound to fail as the cassandra there has been stopped. This only
delays the whole restore step causing un-necessary timoeuts at various
places.

explicitly rely on the StopCassandraStep's result while trying to
exlucde the nodes to be polled from the cluster. By default all
cluster.nodes are to be polled, it is logical to pass a list of nodes
minus the stopped node. And StopCassandraStep conveniently returns the
list of stopped nodes which are considered to be part of the cluster.