use str instead of Path
`pathlib.Path` is a nice API, but the cost of turning all `str` into
`Path` is too high: it's a major cause of memory usage for astacus.

Turn most `Path` into `str`, with a focus on things using in our ipc
models with a large number of objects (manifests, snapshots, etc.)

In most places, this does not change anything. In the few rare case
where we actually do more than comparing paths, use split and join to
do the required operations.

This does not change the storage format or the compatibility with
existing backups: the conversion to `Path` was only an in-memory
behavior of astacus.