don't return result that nobody reads
When starting an operation from the Coordinator to a Node, the Node
was sending an entire result in the result_url that the Coordinator
was not reading at all. It was only using the PUT as a signal to
shorten wait times and actually fetch the result using a GET.

Only PUT a small result object with only the progress, this avoids
serializing and parsing each SnapshotResult twice.