migrate ipc models from pydantic to msgspec
The main goal is to migrate the BackupManifest, but since
everything is connected, this also requires migrating the FastAPI
request and response types (this has a very positive effect on the
memory usage of FastAPI response serialization though).

The config is still using pydantic, to avoid messing with rohmu
models, but that required splitting the Retention model into
one for ipc and one for configuration.