Fix version generation and allow manual setting
Some issues where found with version generation; version didn't follow
pep-440 standards, and it relied on git for actual fetching, while we
may want to override the version from outside to match the actual
version from the rpm package we build.

Also, we don't want generation to happen automagically, but
intentionally, so we had clear targets for what should be happening.