Add option to ignore imbalanced AZs on node map step
When restoring a cluster, the new nodes AZ distribution may not match the one
stored in the backup manifest. In this case, the `MapNode` fails even though the
new nodes may actually be better distributed now.

The change allows mapping new nodes to their backup index in the presence of
imperfect match between node AZs and backup AZs. A new exception is added to
identify the imbalance error and catch it.

The option to the step defaults to false so as to retain backwards compatibility
with the previous behaviour.

Ideally, a better solution would be to consider an AZ imbalance metric relative
to some target replication factor, instead of naively remapping the backup
indices. However, let's say this is good enough for the moment.