clickhouse: make object storage steps sync
The main goal is to migrate `DeleteDanglingObjectStorageFilesStep` to
run in a separate thread instead of keeping astacus stuck while the
step is iterating through very long lists of object storage files.

`RestoreObjectStorageFilesStep` was also migrated because it used
the same APIs as `DeleteDanglingObjectStorageFilesStep`, to avoid
keeping two implementations of everything.