descope fixtures from module to function
Due to changes in pytest-asyncio 0.23, the event loop used by
fixtures/tests with different scopes is not the same. For example,
a fixture declared with "module" scope will use a different event loop
to tests, which run with the "function" scope's event loop. This causes
many tests to fail.
Despite being less efficient, the prudent way forward for now is to
modify the fixtures to be recreated for each test. Astacus' tests are
very fast, so this does not slow down CI runs very much at all.